osx build ctest categories

+++ chaos/file/zppm.cpp
@@ -18,7 +18,7 @@ bool ZPPM::write(ZPath path){
     if(file == NULL)
         return false;
 
-    fprintf(file, "P6\n%lu %lu\n255\n", image.width(), image.height());
+    fprintf(file, "P6\n%lu %lu\n255\n", (unsigned long)image.width(), (unsigned long)image.height());
 
     for(zu64 y = 0; y < image.height(); ++y)
         fwrite(image.buffer() + (y * image.rowSize()), 1, image.rowSize(), file);
